Add info to README.md, and create file structure
Add info about what the project is and its point
in README.md.
Create the directories and html files that will be
used.

index.html has been created in the main
directory.
There is now a "recipes" directory.
"recipes" contains frenchtoast.html, pbnoodles.html
and tunapanini.html.
There is also a directory for images, but is not yet
commited due to being empty.